Introduce deferred status for events
If a property mapping exception is raised during the import, that normally indicates that a reference to an object by remote ID failed to load a local object. Since we aren't able to then ad-hoc fill that object (we need the event data for that) we instead defer the event and simply re-process all deferred events in a second loop.

Events can be repeatedly deferred and there is no limit to the number of such events. This ensures that if a property mapping config error is made, events will keep deferring until the error is fixed, at which point the entire spool is replayed in sequence.