Guard against creating objects with title as UUID
In some cases relationships are returned from the API as (an array of) titles rather than (an array of) UUID vaues. Due to how the typeconverters automatically create objects as placeholders with only the UUID value, when an object doesn't exist, the type converters were capable of creating invalid records which used the title as if it were the UUID.

This added guard defers the event when this happens, until the relation can be resolved.