Forcibly remove all *:n relations before import
Unfortunately, due to the way Extbase persistence works
with references that get removed, deletion of the M:N
record does not always occur. This means that there can
be leftover relations if an entity was previously imported
with more/different relations and then imported again with
new/updated relations. Most prominently a concern with
sys_file_reference which masquerades as M:N but is not
really an M:N with cascading flushes.