[BUGFIX] Prevent overwriting already resolved dimension-data
Fixes a problem where the fallback lookup (which happens if no dimensions were matched or none were defined in TYPO3) happens in too many cases. An additional check to guard if a value was already resolved has been added, along with a comment about a coming refactoring of the fallback value resolving.