[FEATURE] Perform sanitiy check and defer asset deletion if references exist
This patch prevents assets from being deleted if there are any references to the imported asset from any record in the system.

The deletion event gets deferred in this case, and will be retried continually.

Needed solely because PIM events do not cascade from DAM deletions to objects referring to the asset, which means that it is not possible to expect followup events that remove the relation, to be received.