[TASK] Disable reporting unresolvable child as mapping error
Still log the issue, but do not cause deferral for this particular
type of mapping problem. Done in expectation that PIM will
begin to deliver all data regardless of "approval" status which
should remove the false positive deferrals.

At that time the reporting can be re-enabled but until that time,
this workaround prevents needlessly hammering event retries.