[TASK] Improve model and schema generator
Improves the model and schema generation to be more
predictable about the property names resulting from
a given connector configuration.

Due to restructuring in the API responses the previously
used response properties are no longer trustworthy,
since in the case of relations the "name" property now
contains a different value.

Changed so that the index from the field list is used instead,
which matches the field names that are used in responses
for calls to getBeans.

Also implements skipping of fields which will respect the
configured mapping instructions so the connector no longer
generates properties and schema for the properties which
are configured as ignored.